# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP proxy с управлением доступом по IP (whitelist/blacklist/graylist), rate limiting (RPS/RPM/RPH/RPD/bandwidth/соединения) и мониторингом в реальном времени.",
        "title": "Proxy API",
        "contact": {
            "name": "API Support",
            "email": "artiomk00o@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/ip/blacklist": {
            "get": {
                "description": "Возвращает все записи из выбранного списка (IP-адреса, CIDR-подсети, диапазоны).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Получить IP-список",
                "responses": {
                    "200": {
                        "description": "Карта записей: ключ — IP/CIDR/диапазон, значение — true",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет IP-адрес, CIDR-подсеть (например 10.0.0.0/8) или диапазон (192.168.1.1-192.168.1.254) в список. Дубликаты игнорируются.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Добавить запись в IP-список",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись добавлена, false — уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP/CIDR/диапазон или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет IP-адрес, CIDR-подсеть или диапазон из списка. Возвращает false, если записи не было.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Удалить запись из IP-списка",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись удалена, false — не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ip/check": {
            "get": {
                "description": "Определяет, в каком списке находится IP-адрес (black/gray/white/none).\nПорядок проверки: blacklist → graylist → whitelist → none.\nIP можно передать через query-параметр, тело запроса (JSON-строка) или он будет взят из RemoteAddr.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Проверить принадлежность IP к спискам",
                "parameters": [
                    {
                        "type": "string",
                        "example": "192.168.1.1",
                        "description": "Проверяемый IP-адрес",
                        "name": "ip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат проверки",
                        "schema": {
                            "$ref": "#/definitions/handlers.CheckIpResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидный IP-адрес",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ip/graylist": {
            "get": {
                "description": "Возвращает все записи из выбранного списка (IP-адреса, CIDR-подсети, диапазоны).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Получить IP-список",
                "responses": {
                    "200": {
                        "description": "Карта записей: ключ — IP/CIDR/диапазон, значение — true",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет IP-адрес, CIDR-подсеть (например 10.0.0.0/8) или диапазон (192.168.1.1-192.168.1.254) в список. Дубликаты игнорируются.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Добавить запись в IP-список",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись добавлена, false — уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP/CIDR/диапазон или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет IP-адрес, CIDR-подсеть или диапазон из списка. Возвращает false, если записи не было.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Удалить запись из IP-списка",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись удалена, false — не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ip/whitelist": {
            "get": {
                "description": "Возвращает все записи из выбранного списка (IP-адреса, CIDR-подсети, диапазоны).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Получить IP-список",
                "responses": {
                    "200": {
                        "description": "Карта записей: ключ — IP/CIDR/диапазон, значение — true",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет IP-адрес, CIDR-подсеть (например 10.0.0.0/8) или диапазон (192.168.1.1-192.168.1.254) в список. Дубликаты игнорируются.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Добавить запись в IP-список",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись добавлена, false — уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP/CIDR/диапазон или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет IP-адрес, CIDR-подсеть или диапазон из списка. Возвращает false, если записи не было.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-управление"
                ],
                "summary": "Удалить запись из IP-списка",
                "parameters": [
                    {
                        "example": "\"192.168.1.0/24\"",
                        "description": "IP-адрес, CIDR или диапазон",
                        "name": "ip",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true — запись удалена, false — не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный IP или отсутствует тело запроса",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ratelimit": {
            "get": {
                "description": "GET — возвращает все правила. POST — создаёт новое правило для IP или подсети.\n\nПоля правила (все лимиты -1 = отключён):\n- `target`: \"ip\" или \"subnet\"\n- `value`: конкретный IP или CIDR (например 10.0.0.0/8)\n- `max_rps` / `max_rpm` / `max_rph` / `max_rpd`: лимиты запросов\n- `max_upload_bps` / `max_download_bps`: лимиты пропускной способности (байт/с)\n- `max_traffic_day`: дневной лимит трафика (байт)\n- `max_concurrent`: макс. одновременных соединений\n- `max_new_per_second`: макс. новых соединений в секунду",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rate Limiting"
                ],
                "summary": "Управление правилами rate limiting",
                "parameters": [
                    {
                        "description": "Правило rate limiting (только для POST)",
                        "name": "rule",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/domain.RateRule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список всех правил",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.RateRule"
                            }
                        }
                    },
                    "201": {
                        "description": "Правило создано: {\\\"status\\\":\\\"ok\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный JSON или отсутствует поле value/target",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "GET — возвращает все правила. POST — создаёт новое правило для IP или подсети.\n\nПоля правила (все лимиты -1 = отключён):\n- `target`: \"ip\" или \"subnet\"\n- `value`: конкретный IP или CIDR (например 10.0.0.0/8)\n- `max_rps` / `max_rpm` / `max_rph` / `max_rpd`: лимиты запросов\n- `max_upload_bps` / `max_download_bps`: лимиты пропускной способности (байт/с)\n- `max_traffic_day`: дневной лимит трафика (байт)\n- `max_concurrent`: макс. одновременных соединений\n- `max_new_per_second`: макс. новых соединений в секунду",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rate Limiting"
                ],
                "summary": "Управление правилами rate limiting",
                "parameters": [
                    {
                        "description": "Правило rate limiting (только для POST)",
                        "name": "rule",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/domain.RateRule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список всех правил",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.RateRule"
                            }
                        }
                    },
                    "201": {
                        "description": "Правило создано: {\\\"status\\\":\\\"ok\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный JSON или отсутствует поле value/target",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ratelimit/violators/{ip}/unban": {
            "post": {
                "description": "Досрочно снимает блокировку с IP-адреса, нарушившего rate limit. IP передаётся в пути URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rate Limiting"
                ],
                "summary": "Разблокировать IP",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1.2.3.4",
                        "description": "IP-адрес для разблокировки",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\\\"status\\\":\\\"unbanned\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "IP не указан в пути",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Метод не поддерживается",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/top-clients": {
            "get": {
                "description": "Возвращает список клиентских IP, отсортированных по выбранному критерию. Данные накапливаются через middleware с момента старта.\n\nПараметр `sort`:\n- `requests` (по умолчанию) — по количеству запросов\n- `traffic` — по объёму трафика (МБ)\n- `errors` — по проценту ошибок\n\nПоля ответа:\n- `ip`: IP-адрес клиента\n- `requests`: суммарное количество запросов\n- `traffic_mb`: суммарный трафик (МБ)\n- `avg_latency_ms`: средняя задержка (мс)\n- `error_rate`: процент ответов 4xx/5xx\n- `last_seen`: время последнего запроса (HH:MM:SS)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Мониторинг"
                ],
                "summary": "Топ клиентов",
                "parameters": [
                    {
                        "enum": [
                            "requests",
                            "traffic",
                            "errors"
                        ],
                        "type": "string",
                        "default": "requests",
                        "description": "Поле сортировки: requests, traffic, errors",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Максимальное количество записей",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Топ клиентов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.ClientStat"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидный параметр",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ClientStat": {
            "type": "object",
            "properties": {
                "avg_latency_ms": {
                    "type": "number",
                    "example": 18.3
                },
                "error_rate": {
                    "type": "number",
                    "example": 1.2
                },
                "ip": {
                    "type": "string",
                    "example": "1.2.3.4"
                },
                "last_seen": {
                    "type": "string",
                    "example": "13:05:00"
                },
                "requests": {
                    "type": "integer",
                    "example": 4200
                },
                "traffic_mb": {
                    "type": "number",
                    "example": 12.5
                }
            }
        },
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "domain.RateRule": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "max_concurrent": {
                    "description": "Connection limits (-1 = disabled)",
                    "type": "integer",
                    "example": 100
                },
                "max_download_bps": {
                    "type": "integer",
                    "example": 1048576
                },
                "max_new_per_second": {
                    "type": "integer",
                    "example": 20
                },
                "max_rpd": {
                    "type": "integer",
                    "example": 100000
                },
                "max_rph": {
                    "type": "integer",
                    "example": 10000
                },
                "max_rpm": {
                    "type": "integer",
                    "example": 1000
                },
                "max_rps": {
                    "description": "Request rate limits (-1 = disabled)",
                    "type": "integer",
                    "example": 100
                },
                "max_traffic_day": {
                    "description": "Daily total traffic limit in bytes (-1 = disabled)",
                    "type": "integer",
                    "example": 104857600
                },
                "max_upload_bps": {
                    "description": "Bandwidth limits in bytes/s (-1 = disabled)",
                    "type": "integer",
                    "example": 1048576
                },
                "target": {
                    "enum": [
                        "ip",
                        "subnet"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.RuleTarget"
                        }
                    ],
                    "example": "ip"
                },
                "value": {
                    "type": "string",
                    "example": "192.168.1.1"
                }
            }
        },
        "domain.RuleTarget": {
            "type": "string",
            "enum": [
                "ip",
                "subnet"
            ],
            "x-enum-varnames": [
                "TargetIP",
                "TargetSubnet"
            ]
        },
...
        "handlers.CheckIpResponse": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "list": {
                    "type": "string",
                    "enum": [
                        "white",
                        "black",
                        "gray",
                        "none"
                    ],
                    "example": "white"
                },
                "message": {
                    "type": "string",
                    "example": "доступ разрешён"
                }
            }
        },
...
    }
}